move test utilities to testsuite package, add FakeInputs utility
